chore: new module and repository update
