+++ fedot/core/repository/data/data_operation_repository.json
@@ -602,6 +602,16 @@
       "tags": [
         "imbalanced"
       ]
+    },
+    "bagging": {
+      "meta": "ensemble_",
+      "presets": [
+        "fast_train"
+      ],
+      "tags": [
+        "cat_features",
+        "non_applicable_for_ts"
+      ]
     }
   }
 }
